fixed type missmatches

+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/database/dto/MessageIdDTO.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class MessageIdDTO {
 	
-	private Integer cahtId;
+	private Long cahtId;
 	private Integer lastDeleted;
 	private Integer lastSent;
 }
