+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/database/service/XrpDatabaseOperator.java
@@ -26,6 +26,6 @@ public interface XrpDatabaseOperator {
 	
 	MessageIdDTO getMessageId(Long chatId);
 	MessageIdDTO updateMessageId(Long chatId, Integer lastSent, Integer lastDeleted);
-	Boolean deleteMessageId(Integer chatId);
+	Boolean deleteMessageId(Long chatId);
 
 }
